add historical content, we have a working site :)
this include a simple header in front of all the content file.
hierarchy should be reworked (remove the day from the path)
the title will have to be updated manually in most of the entries.

and then theme work to mimick the last version of the diveintomark blog
navigation mechanism